requirements.txt
#requirements.text

# ---------- Core async / storage ----------
aiohappyeyeballs==2.6.1
aiohttp==3.10.5
aiosignal==1.4.0
aiosqlite==0.20.0
anyio==4.10.0
attrs==25.3.0
frozenlist==1.7.0
multidict==6.6.4
yarl==1.20.1

# ---------- Data / viz ----------
altair==5.5.0
numpy==2.3.2
pandas>=2.2.3,<3
pyarrow==21.0.0
plotly==5.21.0
pydeck==0.9.1

# ---------- Web / http ----------
httpx==0.28.1
httpcore==1.0.9
sniffio==1.3.1
requests>=2.32.3,<3
idna==3.10
charset-normalizer==3.4.3
urllib3==2.5.0
certifi==2025.8.3

# ---------- Streamlit & friends ----------
streamlit>=1.39,<2
protobuf>=5.29,<7
watchdog>=5,<7
tornado==6.5.2
blinker==1.9.0
rich==13.9.4
Pygments==2.19.2
typing-extensions>=4.14,<5
packaging==24.2
Jinja2==3.1.6
MarkupSafe==3.0.2
toml==0.10.2
pillow==10.4.0
cachetools==5.5.0
narwhals>=2.1,<3
propcache==0.3.2
streamlit-autorefresh==1.0.1

# ---------- Solana stack (known-good range for py3.13) ----------
solana>=0.36.7,<0.37
solders>=0.23,<0.27
base58==2.1.1
construct==2.10.68
construct-typing==0.6.2
PyNaCl==1.5.0

# ---------- Packaging / build (EXE) ----------
pyinstaller>=6.15,<7
pyinstaller-hooks-contrib>=2025.8,<2026
altgraph==0.17.4
pefile>=2023.2.7; sys_platform == "win32"
pywin32>=306; sys_platform == "win32"
pywin32-ctypes>=0.2.3; sys_platform == "win32"

# ---------- DB / ORM ----------
SQLAlchemy==2.0.43
greenlet>=3.0,<4

# ---------- Git helpers ----------
GitPython==3.1.45
gitdb==4.0.12
smmap==5.0.2

# ---------- Misc utils ----------
python-dotenv>=1.0.1,<2
PyYAML==6.0.2
python-dateutil==2.9.0.post0
pytz>=2025.2,<2026
tzdata>=2025.2,<2026
six==1.17.0
psutil>=6.0.0,<8
websockets>=15.0.1,<16
h11==0.16.0
jsonschema>=4.25,<5
jsonschema-specifications>=2025.4.1,<2026
referencing==0.36.2
rpds-py>=0.27,<0.28
markdown-it-py==4.0.0
mdurl==0.1.2
platformdirs==4.3.6
jsonalias==0.1.1
streamlit-option-menu==0.3.6
colorama==0.4.6
